moved big2d2f from examples to perftest since unpleasantly large 30GB / >30s runtime
