+++ CHANGELOG
@@ -1,5 +1,6 @@
 List of features / changes made / release notes, in reverse chronological order
 
+* change nj and nk in plan to BIGINT (int64_t), fixing Issue #215.
 * library private namespacing via finufft::* of all symbols that do not already
   begin with finufft{f}. This addresses clash with linking against cufinufft
   (their Issue #138).
@@ -7,7 +8,6 @@ List of features / changes made / release notes, in reverse chronological order
 * spread_opts.h fix duplication for FLT=single/double by making FLT->double.
 * examples/simulplans1d1 demos ability to to wield independent plans.
 * sped up float32 1d type 3 by 20% by using float32 cos()... thanks Wenda Zhou.
-* change nj and nk in plan to BIGINT.
 
 V 2.0.4 (1/13/22)
