+++ makefile
@@ -127,7 +127,7 @@ all: test perftest lib examples fortran matlab octave python
 usage:
 	@echo "Makefile for FINUFFT library. Please specify your task:"
 	@echo " make lib - build the main library (in lib/ and lib-static/)"
-	@echo " make examples - compile and run codes in examples/"
+	@echo " make examples - compile and run all codes in examples/"
 	@echo " make test - compile and run quick math validation tests"
 	@echo " make perftest - compile and run (slower) performance tests"
 	@echo " make fortran - compile and run Fortran tests and examples"
@@ -195,11 +195,10 @@ endif
 
 # examples (C++/C) -----------------------------------------------------------
 # build all examples (single-prec codes separate, and not all have one)...
+EXAMPLES = $(basename $(wildcard examples/*.*))
 # ...except only build threadsafe ones if user switch on (thus FFTW>=3.3.6):
 ifeq (,$(findstring FFTW_PLAN_SAFE,$(CXXFLAGS)))
-  EXAMPLES = $(filter-out %/threadsafe1d1 %/threadsafe2d2f, $(basename $(wildcard examples/*.*)))
-else
-  EXAMPLES = $(basename $(wildcard examples/*.*))
+  EXAMPLES := $(filter-out %/threadsafe1d1 %/threadsafe2d2f, $(EXAMPLES))
 endif
 examples: $(EXAMPLES)
 ifneq ($(MINGW),ON)
@@ -299,9 +298,13 @@ spreadtest: $(ST) $(STF)
 spreadtestall: $(ST) $(STF)
 	(cd perftest; ./spreadtestall.sh)
 
+bigtest: perftest/big2d2f
+	@echo "\nRunning >2^31 size example (takes 30 s and 30 GB RAM)..."
+	perftest/big2d2f
+
 PERFEXECS := $(basename $(wildcard test/finufft?d_test.cpp))
 PERFEXECS += $(PERFEXECS:%=%f)
-perftest: $(ST) $(STF) $(PERFEXECS)
+perftest: $(ST) $(STF) $(PERFEXECS) bigtest
 # here the tee cmd copies output to screen. 2>&1 grabs both stdout and stderr...
 	(cd perftest ;\
 	./spreadtestnd.sh 2>&1 | tee results/spreadtestnd_results.txt ;\
